pyproject.toml
[project]
name = "sandstone-redlining"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.12.0",
    "instructor>=1.11.3",
    "openai>=1.109.1",
    "pydantic>=2.11.10",
    "python-dotenv>=1.1.1",
    "rich>=14.1.0",
    "rouge-score>=0.1.2",
    "sentence-transformers>=5.1.1",
    "striprtf>=0.0.29",
]
